==> ML_UnsupervisedLearning_AcademicSuccess/00_data_cleaning.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

# Estructura de les dades i descriptiva

## Selecció de columnes

Abans de començar a explorar les dades i treballar sobre elles se seleccionen les columnes que més ens interessen de la base de dades escollida pel treball. Les columnes se seleccionen de tal manera que compleixin els requisits plantejats, un mínim de 7 columnes numèriques i un mínim de 7 categòriques, on almenys 2 són binàries. Les columnes amb les quals ens quedem s'han explicat en la secció anterior d'aquest informe.

Per tal de facilitar la interpretació de les dades al llarg del treball, també es recodifiquen les variables categòriques amb el valor que correspon a cada nivell. Per exemple, per a la variable binària *daytime_attendance*, s'assigna el valor *Daytime* als 0's i *Evening* als 1's, d'acord amb el diccionari que acompanyava la base de dades. D'aquesta manera, els diferents nivells de les variables categòriques poden interpretar-se directament.

També s'arrodoneixen a dos decimals els valors de les variables numèriques per facilitar la comprensió i simplificar les dades a l'hora de treballar. Arrodonir valors numèrics no és problema en aquest cas, ja que donada la naturalesa de les variables aquest arrodoniment no provoca cap pèrdua d'informació ni ambigüitat entre valors.

Finalment, com que la base de dades no conté *missing values*, s'afegeixen uns quants artificialment de forma aleatòria per tal de poder treballar les tècniques de *missing values* vistes a classe. Aquests *missing values* simulats es creen proporcionalment i s'ha imposat que suposin un 5% dels valors de la base de dades.

```{r}
# Subset data to keep only relevant columns
df = read.csv("data.csv", header = TRUE)

cols_to_keep = c(
  'Curricular.units.1st.sem..without.evaluations.',
  'Curricular.units.2nd.sem..without.evaluations.',
  'Age.at.enrollment', 'Unemployment.rate', 'GDP',
  'Inflation.rate', 'Nacionality', 'Course',
  'Curricular.units.1st.sem..grade.', 'Gender',
  'Curricular.units.2nd.sem..grade.', 'Target',
  'Father.s.occupation', 'Mother.s.occupation',
  'Previous.qualification', 'Daytime.evening.attendance',
  'Scholarship.holder', 'Educational.special.needs'
)

df_ = df[, names(df) %in% cols_to_keep]
# names(df_)
```

```{r}
# Rename cols
colnames(df_) = c('course', 'daytime_attendance', 'previous_qualification', 'nationality', 
                        'occupation_mother', 'occupation_father', 'special_needs', 'gender', 
                        'scholarship', 'age', 'grades_sem_1', 'no_eval_sem_1', 'grades_sem_2', 
                        'no_eval_sem_2', 'unemployment_rate', 'inflation_rate', 'gdp', 'target')

# Convert categorical cols as factors
cat_cols = c('course', 'daytime_attendance', 'nationality', 'occupation_mother', 'occupation_father',
             'special_needs', 'gender', 'scholarship', 'target', 'previous_qualification')

df_[cat_cols] = lapply(df_[cat_cols], factor)

# Rename factor levels
library(dplyr)

df_$course = recode(df_$course, '1' = 'Biofuel Production Technologies', '2' = 'Animation and Multimedia Design', '3' = 'Social Service (evening)', '4' = 'Agronomy', '5' = 'Communication Design', '6' = 'Veterinary Nursing', '7' = 'Informatics Engineering', '8' = 'Equiniculture', '9' = 'Management', '10' = 'Social Service', '11' = 'Tourism', '12' = 'Nursing', '13' = 'Oral Hygiene', '14' = 'Advertising and Marketing Management', '15' = 'Journalism and Communication', '16' = 'Basic Education', '17' = 'Management (evening)')

df_$daytime_attendance = recode(df_$daytime_attendance, '1' = 'Daytime', '0' = 'Evening')

df_$nationality = recode(df_$nationality, '1' = 'Portuguese', '2' = 'German', '3' = 'Spanish', '4' = 'Italian', '5' = 'Dutch', '6' = 'English', '7' = 'Lithuanian', '8' = 'Angolan', '9' = 'Cape Verdean', '10' = 'Guinean', '11' = 'Mozambican', '12' = 'Santomean', '13' = 'Turkish', '14' = 'Brazilian', '15' = 'Romanian', '16' = 'Moldova', '17' = 'Mexican', '18' = 'Ukranian', '19' = 'Russian', '20' = 'Cuban', '21' = 'Colombian')

df_$occupation_mother = recode(df_$occupation_mother, '1' = 'Student', '2' = 'Executive Manager or Legislative Representative', '3' = 'Intellectual Specialist', '4' = 'Technician or Professor', '5' = 'Administrative Staff', '6' = 'Safety Worker', '7' = 'Agriculture, Fishery and Forestry Worker', '8' = 'Construction', '9' = 'Installation Worker', '10' = 'Unskilled Worker', '11' = 'Army Soldier', '12' = 'Other', '13' = 'Blank', '14' = 'Army Officer', '15' = 'Army Sergeant', '16' = 'Army Other', '17' = 'Administrative Director', '18' = 'Hospitality Director', '19' = 'STEM Specialist', '20' = 'Health Professional', '21' = 'Teacher', '22' = 'Finance Specialist', '23' = 'STEM Intermediate Level', '24' = 'Health Technican', '25' = 'Legal or Cultural Technician', '26' = 'Information and Communication Technician', '27' = 'Office Worker', '28' = 'Accounting or Financial Services', '29' = 'Other Administrative Support Staff', '30' = 'Personal Service Worker', '31' = 'Seller', '32' = 'Personal Care Worker', '33' = 'Security Service Personnel', '34' = 'Skilled Agricultural Worker', '35' = 'Subsistence Hunter and Gatherer', '36' = 'Skilled Construction Worker', '37' = 'Skilled Metallurgy Worker', '38' = 'Skilled Electricity Worker', '39' = 'Craft Worker', '40' = 'Machine Operator', '41' = 'Assembly Worker', '42' = 'Mobile Equipment Operator', '43' = 'Unskilled Agricultural Worker', '44' = 'Unskilled Construction and Manufacturing Worker', '45' = 'Meal Preparation Assistant', '46' = 'Street Vendor')

df_$occupation_father = recode(df_$occupation_father, '1' = 'Student', '2' = 'Executive Manager or Legislative Representative', '3' = 'Intellectual Specialist', '4' = 'Technician or Professor', '5' = 'Administrative Staff', '6' = 'Safety Worker', '7' = 'Agriculture, Fishery and Forestry Worker', '8' = 'Construction', '9' = 'Installation Worker', '10' = 'Unskilled Worker', '11' = 'Army Soldier', '12' = 'Other', '13' = 'Blank', '14' = 'Army Officer', '15' = 'Army Sergeant', '16' = 'Army Other', '17' = 'Administrative Director', '18' = 'Hospitality Director', '19' = 'STEM Specialist', '20' = 'Health Professional', '21' = 'Teacher', '22' = 'Finance Specialist', '23' = 'STEM Intermediate Level', '24' = 'Health Technican', '25' = 'Legal or Cultural Technician', '26' = 'Information and Communication Technician', '27' = 'Office Worker', '28' = 'Accounting or Financial Services', '29' = 'Other Administrative Support Staff', '30' = 'Personal Service Worker', '31' = 'Seller', '32' = 'Personal Care Worker', '33' = 'Security Service Personnel', '34' = 'Skilled Agricultural Worker', '35' = 'Subsistence Hunter and Gatherer', '36' = 'Skilled Construction Worker', '37' = 'Skilled Metallurgy Worker', '38' = 'Skilled Electricity Worker', '39' = 'Craft Worker', '40' = 'Machine Operator', '41' = 'Assembly Worker', '42' = 'Mobile Equipment Operator', '43' = 'Unskilled Agricultural Worker', '44' = 'Unskilled Construction and Manufacturing Worker', '45' = 'Meal Preparation Assistant', '46' = 'Street Vendor') 

df_$special_needs = recode(df_$special_needs, '1' = 'Yes', '0' = 'No')

df_$gender = recode(df_$gender, '1' = 'Male', '0' = 'Female')

df_$previous_qualification = recode(df_$previous_qualification, '1' = 'Secondary education', '2' = "Bachelor's degree", '3' = 'Degree', '4' = "Master's degree", '5' = 'PhD', '6' = 'Frequency of higher education', '7' = '12th year incomplete', '8' = '11th year incomplete', '9' = '11th year other', '10' = '10th year of schooling', '11' = '10th year of schooling incomplete', '12' = 'Basic education 3rd cycle', '13' = 'Basic education 2nd cycle', '14' = 'Technological specialization course', '15' = 'Degree (1st cycle)', '16' = 'Professional higher technical course', '17' = "Master's degree (2nd cycle)")

df_$scholarship = recode(df_$scholarship, '1' = 'Yes', '0' = 'No')

# Round grade columns
df_$grades_sem_1 = round(df_$grades_sem_1, 2)
df_$grades_sem_2 = round(df_$grades_sem_2, 2)

# Artificially add random NAs to our data
raw_data = as.data.frame(lapply(df_, function(cc) cc[sample(c(TRUE, NA), prob = c(0.95, 0.05), size = length(cc), replace = TRUE)]))

save(raw_data, file = 'raw_data.Rdata')
```
