==> ML_UnsupervisedLearning_AcademicSuccess/01_profiling_raw_data.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Anàlisi descriptiva abans del preprocessing

En aquesta secció es fa una anàlisi descriptiva (univariant i bivariant) de les dades crues abans del preprocessing.

```{r}
library(ggplot2)
library(knitr)
library(summarytools)
library(tidyverse)

profiling = function(X, col_name, data){
  if (!(is.numeric(X) || class(X)=="Date")){ 
    freqs = table(as.factor(X), useNA = "no"); 
    proportions = freqs/nrow(data)
    df = data.frame(x = levels(as.factor(X)), y = freqs)
    par(mfrow = c(1,2))
    print(ggplot(df, aes(x = "", y = freqs, fill = x)) + geom_bar(stat = "identity",
    width = 1) + scale_y_continuous(expand = c(0,0)) + coord_polar("y", start = 0) + 
    theme_minimal())
    print(ggplot(df, aes(x = x, y = freqs)) + geom_bar(stat = "identity", color =
    "lightblue", fill = "lightblue") + 
    geom_text(aes(label = x), nudge_y = -5, size = 3) + 
    scale_y_continuous(expand = c(0,0), name = "Frequency") + 
    scale_x_discrete(name = col_name) + 
    theme_minimal() +
    theme(axis.text.x = element_blank(), 
          axis.ticks.x = element_blank(), 
          axis.text.y = element_text(size = 10),
          axis.line.x = element_line(color = "black", size = 0.5),
          panel.grid.major.y = element_line(color = "gray", size = 0.2)))
    x = data.frame(X)
    colnames(x) = col_name
    st_options(plain.ascii  = FALSE,  headings = FALSE, footnote = NA, style = 'grid',
               dfSummary.varnumbers = FALSE, dfSummary.valid.col = FALSE, tmp.img.dir  = "img",
               round.digits = 3, dfSummary.silent=TRUE, dfSummary.graph.col= FALSE)
    dfSummary(x)
    
   }else{
      print(ggplot(data, aes(x = X)) + geom_bar(col = "lightblue", fill = "lightblue") + theme_minimal())
       taula = descr(X, transpose = TRUE, style = "rmarkdown", justify = "center", 
                 stats = c("N.Valid", "min", "q1", "med", "mean", "sd", "q3", "max", "iqr"))
       rownames(taula) = NULL
       print(taula)
   }
}
```

```{r}
load('raw_data.Rdata')
```

### Anàlisi univariant de les variables

**Variables categòriques**

Pel que fa a l'anàlisi descriptiva univariant per a variables categòriques, es creen gràfiques de tipus *pie chart* que representen la part proporcional de cada nivell o categoria sobre el total d'observacions de la variable categòrica així com un gràfic de barres que compta les ocurrències de cada categoria o nivell de la variable. La mateixa informació que mostren aquests dos gràfics es pot representar en una taula de freqüències. 

*COURSE*

```{r, out.width="50%"}
profiling(raw_data$course, 'course', raw_data)
```

El curs amb menys estudiants és el de "Biofuel Production Techno" amb un total de 12 alumnes. Així mateix, el curs amb més estudiants és el de "Nursing".

*DAYTIME_ATTENDANCE*

```{r, out.width="50%"}
profiling(raw_data$daytime_attendance, 'daytime_attendance', raw_data)
```

La major part de l'assistència es concentra en la franja horària del matí.

*NATIONALITY*

```{r, out.width="50%"}
profiling(raw_data$nationality, 'nationality', raw_data)
```

El 97.5% dels alumnes són portuguesos. Les altres 20 nacionalitats es distribueixen entre el 2.5% restant dels individus.

*OCCUPATION_MOTHER*

```{r, out.width="50%"}
profiling(raw_data$occupation_mother, 'occupation_mother', raw_data)
```

Les ocupacions de les mares més destacades són treballadora no qualificada (35.5%) i personal administratiu (18.4%).

*OCCUPATION_FATHER*

```{r, out.width="50%"}
profiling(raw_data$occupation_father, 'occupation_father', raw_data)
```

La feina més representativa dels pares és treballador no qualificat, seguida dels llocs de treball a la construcció.

*SPECIAL_NEEDS*

```{r, out.width="50%"}
profiling(raw_data$special_needs, 'special_needs', raw_data)
```

El 98.9% dels alumnes no tenen necessitats especials.

*GENDER*

```{r, out.width="50%"}
profiling(raw_data$gender, 'gender', raw_data)
```

Aproximadament el 65% dels individus s'identifiquen com a dona.

*SCHOLARSHIP*

```{r, out.width="50%"}
profiling(raw_data$scholarship, 'scholarship', raw_data)
```

Tres quartes parts dels alumnes no reben cap mena d'ajuda econòmica.

*TARGET*

```{r, out.width="50%"}
profiling(raw_data$target, 'target', raw_data)
```

La meitat dels enquestats són graduats, el 32% ha abandonat els estudis i quasi el 18% són persones que han reprès els estudis.

*PREVIOUS_QUALIFICATION*

```{r, out.width="50%"}
profiling(raw_data$previous_qualification, 'previous_qualification', raw_data)
```

El 84% dels estudiants tenen l'Educació Secundària Obligatòria.

**Variables numèriques**

Per a l'anàlisi univariant de variables numèriques s'utilitza un histograma per analitzar visualment els valors que pren aquesta variable. També es fa un resum (i.e. *summary*) de la variable que permet conèixer, entre altres coses, quina és la mitjana, la mediana, el valor màxim i el valor mínim que pren.

*AGE*

```{r, out.width="60%"}
profiling(raw_data$age, 'age', raw_data)
```

La variable edat es comprèn dins un interval que va dels 17 als 70 anys amb ambdós inclosos. No obstant això, la població es distribueix principalment al voltant dels 20 anys.

*GRADES_SEM_1*

```{r, out.width="60%"}
profiling(raw_data$grades_sem_1, 'grades_sem_1', raw_data)
```

Les notes del primer semestre es reparteix en gran manera entre 10 i 14.

*NO_EVAL_SEM_1*

```{r, out.width="60%"}
profiling(raw_data$no_eval_sem_1, 'no_eval_sem_1', raw_data)
```

Majoritàriament, no hi ha unitats curriculars no avaluades el primer semestre.

*GRADES_SEM_2*

```{r, out.width="60%"}
profiling(raw_data$grades_sem_2, 'grades_sem_2', raw_data)
```

Les notes del segon semestre es reparteix en gran manera entre 10 i 14.

*NO_EVAL_SEM_2*

```{r, out.width="60%"}
profiling(raw_data$no_eval_sem_2, 'no_eval_sem_2', raw_data)
```

Majoritàriament, no hi ha unitats no avaluades el segon semestre.

*UNEMPLOYMENT_RATE*

```{r, out.width="60%"}
profiling(raw_data$unemployment_rate, 'unemployment_rate', raw_data)
```

El valor més baix que pren la taxa d'atur és 7.6% i el més alt, 16.2%.

*INFLATION_RATE*

```{r, out.width="60%"}
profiling(raw_data$inflation_rate, 'inflation_rate', raw_data)
```

La màxima taxa d'inflació és del 3.7%.

*GDP*

```{r, out.width="60%"}
profiling(raw_data$gdp, 'gdp', raw_data)
```

El PIB oscil·la entre -4 i 3.5.

### Anàlisi bivariant

Per a l'anàlisi bivariant es miren les combinacions més interessants (i que més sentit té fer) de dues variables. 

**Dues variables numèriques**

A l'hora d'analitzar conjuntament dues variables numèriques, s'utilitza un *scatterplot* on cada variable està representada en un dels dos eixos.

*AGE* i *GRADES_SEM_1*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = age, y = grades_sem_1, color = grades_sem_1)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

Es veu que no hi ha diferències significatives en la nota del primer semestre segons l'edat. Però en les edats més avançades, les notes poden no arribar a ser tan altes.

*AGE* i *GRADES_SEM_2*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = age, y = grades_sem_2, color = grades_sem_2)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

S'observa el mateix que en el gràfic anterior amb les notes del primer semestre.

*AGE* i *NO_EVAL_SEM_1*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = age, y = no_eval_sem_1, color = no_eval_sem_1)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

Segons l'edat, no sembla que hi hagi diferències en les unitats no avaluades.

*AGE* i *NO_EVAL_SEM_2*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = age, y = no_eval_sem_2, color = no_eval_sem_2)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

Ocorre el mateix que en les unitats no avaluades del primer semestre, l'edat no sembla ser significativa.

*GRADES_SEM_1* i *NO_EVAL_SEM_1*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = grades_sem_1, y = no_eval_sem_1, color = no_eval_sem_1)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

Si es mira el gràfic, no s'observa gaire correlació entre les notes i les unitats no avaluades del primer semestre.

*GRADES_SEM_2* i *NO_EVAL_SEM_2*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = grades_sem_2, y = no_eval_sem_2, color = no_eval_sem_2)) +
  geom_jitter()  +
  scale_color_gradient(low = "lightblue", high = "orange") +
  theme_minimal()
```

No s'observa gaire correlació entre les notes i les unitats no avaluades del segon semestre.

**Dues variables categòriques**

Quan hi ha dues variables categòriques, s'utilitza un gràfic de barres per representar les ocurrències d'una de les variables separant entre nivells o categories de l'altra variable.

*NATIONALITY* i *DAYTIME_ATTENDANCE*

Per a aquesta primera anàlisi, en ser la majoria dels alumnes de nacionalitat portuguesa se separen el nombre de casos de persones portugueses respecte als altres casos, ja que si no, no es pot apreciar bé en el gràfic com es distribueixen els casos dels individus no portuguesos.

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$nationality == "Portuguese" | raw_data$nationality == "Unknown",], aes(x = nationality, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))

ggplot(data = raw_data[raw_data$nationality != "Portuguese" & raw_data$nationality != "Unknown",], aes(x = nationality, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Pel cas de persones portugueses, hi ha major assistència durant el matí. Passa el mateix amb els casos de les altres nacionalitats.

*SPECIAL_NEEDS* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = special_needs, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Els individus que no tenen necessitats especials assisteixen majoritàriament al matí, mentre que els que sí que tenen necessitats especials, només assisteixen al matí.

*GENDER* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = gender, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Independentment del gènere, la majoria va a classe en horari de matí.

*SCHOLARSHIP* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = scholarship, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

La majoria de les persones van de matins, sense importar si tenen beca o no.

*PREVIOUS_QUALIFICATION* i *DAYTIME_ATTENDANCE*

Com que la major part dels individus tenen "Secondary education" com a qualificació prèvia, s'han separat aquests casos respecte als altres per tal de poder observar amb millor precisió com es distribueixen.

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$previous_qualification=="Secondary education",], aes(x = previous_qualification, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() 

ggplot(data = raw_data[raw_data$previous_qualification!="Secondary education",], aes(x = previous_qualification, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Sembla que independentment de la qualificació prèvia, els individus assisteixen durant el matí.

*TARGET* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = target, fill = daytime_attendance)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Independentment de si van acabar els estudis, els van abandonar o els van continuar més tard, hi ha una tendència a anar al matí a classes.

*NATIONALITY* i *SPECIAL_NEEDS*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$nationality == "Portuguese" | raw_data$nationality == "Unknown",], aes(x = nationality, fill = special_needs)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal()


ggplot(data = raw_data[raw_data$nationality != "Portuguese" & raw_data$nationality != "Unknown",], aes(x = nationality, fill = special_needs)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Es pot observar que la majoria d'estudiants de nacionalitat portuguesa no tenen necessitats especials. Per a les altres nacionalitats, no es pot saber significativament la probabilitat de tenir necessitats especials, ja que la mida de la mostra és molt petita, però sembla que no tenen necessitats especials.

*NATIONALITY* i *GENDER*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$nationality == "Portuguese" | raw_data$nationality == "Unknown",], aes(x = nationality, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal()

ggplot(data = raw_data[raw_data$nationality != "Portuguese" & raw_data$nationality != "Unknown",], aes(x = nationality, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

S'observa més tendència a ser estudiant dona que home tant en la nacionalitat portuguesa com en les altres.

*NATIONALITY* i *SCHOLARSHIP*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$nationality == "Portuguese" | raw_data$nationality == "Unknown",], aes(x = nationality, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal()

ggplot(data = raw_data[raw_data$nationality != "Portuguese" & raw_data$nationality != "Unknown",], aes(x = nationality, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Per a les persones portugueses, la probabilitat de tenir beca és aproximadament un quart. Pels casos de nacionalitat desconeguda s'observa una tendència similar.

*NATIONALITY* i *TARGET*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$nationality == "Portuguese" | raw_data$nationality == "Unknown",], aes(x = nationality, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal()

ggplot(data = raw_data[raw_data$nationality != "Portuguese" & raw_data$nationality != "Unknown",], aes(x = nationality, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

La probabilitat de graduar-se en persones portugueses és més gran que la de deixar la carrera. Ocorre una tendència similar amb les altres nacionalitats.

*SPECIAL_NEEDS* i *GENDER*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = special_needs, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

El gènere no sembla rellevant respecte de tenir necessitats especials o no.

*SPECIAL_NEEDS* i *SCHOLARSHIP*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = special_needs, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Sembla que en tots els casos, tenint o no necessitats especials, la probabilitat de no rebre una beca és més gran que la probabilitat que la concedeixin.

*TARGET* i *SPECIAL_NEEDS*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = target, fill = special_needs)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Per a les persones amb necessitats especials s'observa la mateixa tendència que sense necessitats especials, és més probable graduar-se que deixar la carrera.

*PREVIOUS_QUALIFICATION* i *GENDER*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$previous_qualification=="Secondary education",], aes(x = previous_qualification, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal()

ggplot(data = raw_data[raw_data$previous_qualification!="Secondary education",], aes(x = previous_qualification, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Pel que fa a les persones amb educació secundària com a qualificació prèvia, dos terços són dones.

Si s'observen com es distribueix el gènere en les altres qualificacions prèvies, no ocorre el mateix que per a l'educació secundària. Per exemple, hi ha més persones amb un curs d'especialització tecnològica que són homes que dones. No obstant això, per a aquests casos la mida mostral no és gaire gran.

*GENDER* i *SCHOLARSHIP*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = gender, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

És més probable no tenir "scholarship" que tenir-ne, tant en el cas dels homes com en el de les dones. Tot i això, és més probable tenir beca en el cas de les dones.

*TARGET* i *GENDER*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = target, fill = gender)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

Tenint en compte que hi ha més dones que homes inscrits, la probabilitat de deixar la carrera és més alta com a home.

*PREVIOUS_QUALIFICATION* i *SCHOLARSHIP*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$previous_qualification=="Secondary education",], aes(x = previous_qualification, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal()

ggplot(data = raw_data[raw_data$previous_qualification!="Secondary education",], aes(x = previous_qualification, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple", "orange")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Per a la majoria de les qualificacions prèvies, la probabilitat de tenir beca sembla inferior que la de no tenir-ne.

*TARGET* i *SCHOLARSHIP*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = target, fill = scholarship)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue","purple")) + theme_minimal()
```

La majoria de les persones amb beca s'han graduat. Les persones sense beca tenen una probabilitat similar de deixar-ho que de graduar-se, aquesta última sent una mica més alta.

*PREVIOUS_QUALIFICATION* i *TARGET*

```{r, out.width="50%"}
ggplot(data = raw_data[raw_data$previous_qualification=="Secondary education",], aes(x = previous_qualification, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal()

ggplot(data = raw_data[raw_data$previous_qualification!="Secondary education",], aes(x = previous_qualification, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

Pels individus que tenen com a qualificació prèvia "Secondary education" tenen més probabilitats de graduar-se. No obstant això, per als altres casos la distribució canvia i, per exemple, si tenen "Basic education 3rd cycle" hi ha aproximadament el doble de probabilitats d'abandonar la carrera que de graduar-se.

*OCCUPATION_MOTHER* i *TARGET*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = occupation_mother, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

L'ocupació de la mare no sembla una variable rellevant per a si la persona es gradua o no.

*OCCUPATION_FATHER* i *TARGET*

```{r, out.width="60%"}
ggplot(data = raw_data, aes(x = occupation_father, fill = target)) +
    geom_bar(position = "dodge") +  scale_fill_manual(values= c("lightblue", "purple", "orange", "red")) + theme_minimal() + theme(axis.text.x = element_text(angle = 90, vjust = 0.5))
```

L'ocupació del pare no sembla una variable rellevant per a si la persona es gradua o no.

**Una variable numèrica i una de categòrica**

Quan hi ha una variable numèrica i una variable categòrica es fa l'anàlisi multivariant de la mateixa manera que en el cas anterior. La variable numèrica és representada en l'eix de les X i per cada valor d'aquesta variable es representa en format gràfic de barres les ocurrències de cada categoria o nivell de la variable categòrica.

*AGE* i *GENDER*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = age, fill = gender)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange")) 
```

Hi ha més tendència a estudiar en edats joves i en dones. A partir dels 25 anys aproximadament, hi ha un percentatge similar de dones que d'homes que estudien.

*AGE* i *SCHOLARSHIP*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = age, fill = scholarship)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange")) 
```

Les beques es concedeixen en edats més joves.

*AGE* i *SPECIAL_NEEDS*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = age, fill = special_needs)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange")) 
```

Pel que fa a les necessitats especials, no sembla que estigui correlacionat amb l'edat.

*AGE* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = age, fill = daytime_attendance)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange")) 
```

Amb l'edat, incrementa el percentatge d'assistència a les tardes. No obstant això, continua dominant l'assistència al matí en tots els grups d'edats.

*AGE* i *TARGET*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = age, fill = target)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange")) 
```

A mesura que avança l'edat, incrementa la probabilitat relativa a deixar la carrera.

*GRADES_SEM_1* i *GENDER*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = grades_sem_1, fill = gender)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```

Les notes del primer semestre semblen distribuir-se igualment entre els dos gèneres. Sabent que el 0 significa no presentat s'observa que els homes tenen més tendència a no presentar-se que les dones, tenint en compte que hi ha més dones que homes.

*GRADES_SEM_1* i *SCHOLARSHIP*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = grades_sem_1, fill = scholarship)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```

Hi ha correlació entre treure notes més altes i tenir una "scholarship". Si no es té beca és més probable no presentar-se.

*GRADES_SEM_1* i *DAYTIME_ATTENDANCE*

```{r, out.width="60%"}
ggplot(raw_data, aes(x = grades_sem_1, fill = daytime_attendance)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```

Les persones que assisteixen a classes de tarda tenen una mitjana de notes en el primer semestre més baixa que les persones que van a classes al matí.

Referent a les notes del segon semestre, es repliquen els mateixos patrons que per a les notes del primer.

*GRADES_SEM_2* i *GENDER*

```{r, out.width="59%"}
ggplot(raw_data, aes(x = grades_sem_2, fill = gender)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```

*GRADES_SEM_2* i *SCHOLARSHIP*

```{r, out.width="59%"}
ggplot(raw_data, aes(x = grades_sem_2, fill = scholarship)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```

*GRADES_SEM_2* i *DAYTIME_ATTENDANCE*

```{r, out.width="59%"}
ggplot(raw_data, aes(x = grades_sem_2, fill = daytime_attendance)) +
    geom_histogram(color="#e9ecef", alpha=1, position = 'dodge', binwidth = 2) +
    scale_fill_manual(values = c("lightblue", "purple", "orange"))
```
