==> ML_UnsupervisedLearning_AcademicSuccess/02_prepro.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Preprocessing

En la primera part del preprocessing s'utilitzaran diferents mètodes per detectar outliers univariants i multivariants. En la segona part es tractaran els valors faltants de les variables categòriques i numèriques. L'objectiu és netejar la base de dades per poder garantir que les següents anàlisis siguin acurades.

```{r, echo = FALSE}
library(assertr)
library(tidyverse)
library(magrittr)
```

```{r, echo = FALSE}
# Carregar les dades raw_data
load("raw_data.Rdata")
data_nona <- na.omit(raw_data)
```

### Detecció d'outliers

Es duu a terme una anàlisi per detectar outliers tant univariant com multivariant, ja que els valors atípics poden afectar potencialment a l'estimació dels paràmetres i a les conclusions extretes de l'informe. 

**Outliers multivariants**

Per a fer la detecció d'outliers multivariants s'utilitzen les dades amb només variables numèriques i sense valors nuls, pel fet que es necessiten aquest tipus de dades per a dur a terme el mètode escollit: Distància de Mahalanobis. Aquesta mesura serveix per mesurar la distància entre dos vectors, on cada vector és una observació (individu). La distància de Mahalanobis es basa en la distància euclidiana i permet identificar combinacions atípiques de diverses variables.

````{r, echo = FALSE}
nums <- unlist(lapply(data_nona, is.numeric), use.names = FALSE)  
data_num_nona <- data_nona[ , nums]
# Mahalanobis distance for each row of data frame
data_num_nona['mahala'] <- maha_dist(data_num_nona)
# Critical values for eight degrees of freedom of chi square 
chiq <- qchisq(0.99, 8) 
# > 18.47531
# data_num_nona[data_num_nona$mahala > chiq, ]
dim <- dim(data_num_nona[data_num_nona$mahala > chiq, ])
````

El resultat de l'anàlisi d'outliers multivariant és que hi ha `r dim[1]` files que en contenen algun. Com no es pot reconèixer quines són les variables d'aquestes files que provoquen el valor atípic, es decideix que les dades i files es queden tal com estan perquè convertir les `r dim[1]` files és una gran pèrdua d'informació i podríem incórrer en errors en les conclusions.

**Outliers univariants de les variables numèriques**

- *VARIABLES 'gdp', 'inflation_rate' I 'unemployment_rate'*

````{r, echo = FALSE}
rw <- raw_data
# hist(rw$gdp)
# hist(rw$inflation_rate)
# hist(rw$unemployment_rate)
````

Aquestes tres variables mostren dades macroeconòmiques dels anys en què van estudiar els alumnes. Com que la recollida de dades es va fer en deu anys, els valors s'han d'agrupar en 10 blocs corresponents als anys. No hi ha cap valor que quedi fora d'aquests blocs, per tant, es pot dir que no hi ha valors atípics.

- VARIABLE 'age'

````{r, echo = FALSE}
# summary(rw$age)
# hist(rw$age)
# boxplot(rw$age, 
#         main = "AGE",
#         boxwex = 0.5,col="blue")
````

La variable de l'edat segueix una distribució exponencial. Es poden observar alguns valors atípics, però aquests no responen a errors en la recollida de dades, sinó a casos d'estudiants d'edat més avançada. No hi ha indicis que es tracti de cap error, ja que el valor màxim és de 70, una edat poc comuna però factible per a un estudiant. Així doncs, no és necessari eliminar aquests valors.

- VARIABLES 'grades_sem_1' I 'grades_sem_2'

````{r, echo = FALSE}
# summary(rw$grades_sem_1)
# summary(rw$grades_sem_2)
````

Aquestes dues variables indiquen les qualificacions dels alumnes. En el sistema universitari portuguès (on es desenvolupa l'estudi) les assignatures es qualifiquen del 0 al 20. Observant el rang de les dades es pot veure que no hi ha cap outlier.

- VARIABLES 'no_eval_sem_1' I 'no_eval_sem_2'

````{r, echo = FALSE}
# summary(rw$no_eval_sem_1)
# hist(rw$no_eval_sem_1)
# summary(rw$no_eval_sem_2)
# hist(rw$no_eval_sem_2)
````

Aquestes variables indiquen la quantitat de crèdits no avaluats de cada alumne. Mentre el valor sigui inferior al nombre de crèdits matriculats no hi ha indicis que hi hagi outliers.

### Tractament de NA's

El tractament de NA's es duu a terme per variables categòriques i per variables numèriques. Per les variables categòriques es crea una categoria per agrupar els NA's, mentre que per les variables numèriques s'utilitzarà el mètode d'imputació MIMMI.

**NA's en variables categòriques**

Es crea una variable 'Unknown', que serà un nou nivell de les variables factors, que agrupa tots els NA's d'aquestes variables.

````{r, echo = FALSE}
# varCat <- sapply(rw, is.factor)
# varCat <- names(varCat)[varCat == TRUE]
# 
# for (vC in varCat) {
#   rw[, vC] <- as.character(rw[, vC])
#   quienes <- which(is.na(rw[, vC]))
#   rw[quienes, vC] <- "Unknown"
#   rw[, vC] <- as.factor(rw[, vC])
#   print(levels(rw[ ,vC]))
# }
````

**NA's en variables numèriques (MIMMI)**

Per tractar els valors faltants de les variables numèriques es farà servir el mètode Mixed Intelligent Multivariate Missing Imputation (MIMMI). Aquest mètode permet imputar els NA's en bases de dades multivariants.

Per aplicar el mètode de MIMMI a les dades s'ha treballat sobre el codi de Karina Gibert: *Gibert, K. (2014). Mixed intelligent-multivariate missing imputation. International Journal of Computer Mathematics, 91(1), 85-96 Barcelona May 2019*

````{r, echo = FALSE}
# install.packages("StatMatch")
library(cluster)
require(StatMatch)

# assume missings represented with NA
uncompleteVar <- function(vector){any(is.na(vector))}

Mode <- function(x) 
{
  x <- as.factor(x)
  maxV <- which.max(table(x))
  return(levels(x)[maxV])
}
````

Diem que k=3 perquè la població es divideix en 3 categories (Graduate, Dropout, Enrolled).

````{r, echo = FALSE}
# MiMMi <- function(data, priork=-1)
# {
#   # Identify columns without missings
#   colsMiss <- which(sapply(data, uncompleteVar))
#   if(length(colsMiss) == 0){
#     print("Non missing values found")
#     out <- dd
#   } else {
#     K <- dim(data)[2]
#     colsNoMiss <- setdiff(c(1:K), as.vector(colsMiss))
#     
#     #cluster with complete data
#     dissimMatrix <- daisy(data[ , colsNoMiss], metric = "gower", stand = TRUE)
#     distMatrix <- dissimMatrix^2
#     
#     hcdata <- hclust(distMatrix, method = "ward.D2")
#     plot(hcdata)
#     
#     if(priork == -1){
#       nk <- readline("See the dendrogramm and enter a high number of clusters (must be a positive integer). k: ")
#       nk <- as.integer(nk)
#     } else {nk <- priork}
#     
#     partition <- cutree(hcdata, nk)
#     
#     CompleteData <- data
#     # només cal per tenir traça de com s'ha fet la substitució
#     newCol <- K+1
#     CompleteData[ , newCol] <- partition
#     names(CompleteData)[newCol] <- "ClassAux"
#     
#     setOfClasses <- as.numeric(levels(as.factor(partition)))
#     imputationTable <- data.frame(row.names = setOfClasses)
#     p <- 1
#     
#     for(k in colsMiss)
#     {
#       # Files amb valors utils
#       rowsWithFullValues <- !is.na(CompleteData[,k])
#       
#       # Calcular valors d'imputació
#       if(is.numeric(CompleteData[,k]))
#       {
#         imputingValues <- aggregate(CompleteData[rowsWithFullValues,k], by = list(partition[rowsWithFullValues]), FUN = mean)
#       } else {
#         imputingValues <- aggregate(CompleteData[rowsWithFullValues,k], by = list(partition[rowsWithFullValues]), FUN = Mode)
#       }
#       
#       # Impute
#       for(c in setOfClasses)
#       {
#         data[is.na(CompleteData[,k]) & partition == c,k] <- round(imputingValues[c,2],0)
#       }
#       
#       # Imputation Table
#       imputationTable[,p] <- imputingValues[,2]
#       names(imputationTable)[p] <- names(data)[k]
#       p <- p+1
#     }
#     
#     rownames(imputationTable) <- paste0("c", 1:nk)
#     out <- new.env()
#     out$imputedData <-data
#     out$imputation <- imputationTable
#   }
#   return(out)
# }
# 
# # run MIMMI
# dimpute <-MiMMi(rw)
# 
# # table of imputation values used
# dimpute$imputation
# 
# # imputed dataset
# prepro_data <- dimpute$imputedData
````

````{r, echo = FALSE}
# Es guarda dataframe de dades després del preprocessing
# save(prepro_data, file = 'prepro_data.Rdata')
````
