==> ML_UnsupervisedLearning_AcademicSuccess/07_ACM.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

# ACM

```{r, include=FALSE}
library(FactoMineR)
library(Matrix)
library(ggplot2)
```

```{r, include=FALSE}
# Lectura de la base de dades
bd <- get(load('data_post_pca.RData'))
```

Per començar s'han d'escollir quines variables seran les variables actives, i quines les complementàries:

**Categòriques actives**:

- scolarship

- target

- gender

- daytime_attendance

- occupation_mother

- course

**Categòriques suplementàries**:

- Nationality

- Special_needs

- occupation_father

- previous_qualification

**Numèriques suplementàries**:

Com en l'ACP es va decidir no eliminar cap variable numèrica, s'utilitzaran totes les variables numèriques com a variables suplementàries.

A partir d’aquesta selecció es procedirà a realitzar l’ACM segons el mètode de Burt. Es comença quantificant la inèrcia de cada dimensió a partir de la variància explicada de cada una.

```{r, out.width="95%"}
res.mca <- MCA(bd, quanti.sup = c(10:17), quali.sup = c(3,4,6,7), method = "Burt")
res.mca
```

En els gràfics obtinguts es pot veure la variabilitat que expressen cada una de les variables categòriques en funció de les dimensions 1 i 2. Aquelles variables que estiguin més a prop de l'origen de coordenades aporten molt poca informació respecte a la variabilitat de les dades i, per tant, són poc importants. En canvi, aquelles variables més allunyades del centre aporten informació més rellevant, com l'horari de classes o el curs. 

Es representen gràficament la inèrcia que explica cada una de les dimensions generades:

```{r, out.width="90%"}
barp <- barplot(res.mca$eig[,2], names.arg=1:nrow(res.mca$eig), ylim = c(0,21), 
                cex.names = 0.7, las = 2, xlab = c("Dimensions"), 
                ylab = c("Percentatge de variància explicada"), 
                main = "Inèrcia explicada per cada dimensió")
```

Si una dimensió té una inèrcia baixa, significa que totes les modalitats estan molt properes al centre de gravetat, i en conseqüència, són molt similars. A mesura que augmenta la inèrcia, va augmentant la distància al centre de gravetat i, per tant, es redueix la similitud.

Per poder estudiar-ho més a fons, es realitza la següent taula on es pot observar per a cada dimensió, el seu valor propi, el percentatge de variància (o inèrcia) explicada, i el percentatge de variància (o inèrcia) acumulada:

```{r}
round(res.mca$eig, 2)
```

Tenim un total de 23 dimensions. Es veu que la dimensió 1 és la que destaca més per sobre la resta, explicant un 14.90% de la variabilitat de les dades, seguida de les dimensions 2, que també destaca per sobre de la resta, explicant un 9.34% de la variabilitat de les dades. A partir de la dimensió 3, es veu que el gràfic s'estabilitza bastant ja fins a l'última dimensió. 

Per tant, en total les dues primeres dimensions ja expliquen un 24.25% de la variabilitat de les dades, i es necessiten 15 dimensions per arribar a tenir una inèrcia acumulada per sobre del 80%.

## Plot individus

Es representa gràficament com es distribueixen els individus en funció de les dues primeres dimensions que expliquen un 24.25% de la variabilitat:

```{r, out.width="95%"}
plot(res.mca, invisible = c("var","quali.sup"), cex=0.7)
```

A simple vista, es pot veure que els individus es diferencien en 3 grups. Tot i això, la gran majoria d'individus es concentren molt propers a l'origen de manera que aporten poca informació sobre quines variables rellevants tenen respecte als altres individus. Es passa a estudiar cada variable per observar si hi ha algun tipus d'associació entre elles.

## Plot variables

```{r}
# Totes les variables actives
newbd = bd[, c("target", "daytime_attendance","course","scholarship","gender", "occupation_mother")]
res.mca1<-MCA(newbd, method="Burt", graph=FALSE)
cats = apply(newbd, 2, function(x) nlevels(as.factor(x)))

mca1_vars_df = data.frame(res.mca1$var$coord, Variable = rep(names(cats), cats))
mca1_obs_df = data.frame(res.mca1$ind$coord)

ggplot(data = mca1_obs_df, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = mca1_vars_df, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(mca1_vars_df), colour = Variable), cex = 3) +
  
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable")+ 
  xlim(-1,2.2) + 
  ylim(-1.5,2)
```

Igual que en el gràfic per individus, es veu que les dades es divideixen en 3 grups. Es poden classificar segons els eixos com:

- Grup 1: X = [-1, 0.5], Y = [-1, 1]

- Grup 2: X = [0, 1], Y = [-0.5, 1]

- Grup 3: X = [1, 2], Y = [0.5, 2]

Amb totes les variables actives juntes, és impossible diferenciar-les i poder extreure'n conclusions, per això s'aniran fent gràfics de les diferents variables:

```{r, out.width="90%"}
# variable Scholarship
v1 = c(16:18)
data = mca1_vars_df[v1,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

A simple vista, es pot observar que totes les modalitats de la variable *Scholarship* no estan molt repartides en el gràfic. Es pot veure que les modalitats 'Unknown' i 'No' no expliquen gaire variabilitat en situar-se molt a prop del centre de gravetat de les dimensions. De totes maneres, la categoria 'Yes' sí que sembla que podria tenir certa rellevància explicativa (II quadrant).

```{r, out.width="90%"}
# variable Gender
v2 = c(19:21)
data = mca1_vars_df[v2,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

Com en la variable Scholarship, en la variable *Gender* s'observa que totes les seves modalitats queden agrupades al voltant de l'origen de coordenades. En aquest cas les modalitats 'Female' i 'Unknown' no aporten informació que pugui ser útil, mentre que la categoria 'Male' queda una mica més lluny (quadrant IV) i sembla que podria tenir certa associació.

```{r, out.width="90%"}
# variable daytime_attendance
v3 = c(5:7)
data = mca1_vars_df[v3,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

Pel que fa a la variable *daytime_attendance* es veu que les modalitats es reparteixen d'una forma força interessant en el gràfic. Si bé les modalitats 'Daytime' i 'Unknown' expliquen poca variabilitat en quedar a prop del centre, la categoria 'Evening' queda més separada i se situa en el quadrant I de forma clara.

```{r, out.width="90%"}
# variable Target
v4 = c(1:4)
data = mca1_vars_df[v4,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

Altra vegada, es veu que totes les modalitats de la variable *Target* es concentren al voltant del centre i sense gaire variabilitat. Tot i això, hi ha dues modalitats que sembla que podrien aportar certa informació. La categoria 'Dropout' se situa de forma més clara al quadrant IV i la categoria 'Graduate' se situa al quadrant II.

```{r, out.width="90%"}
# variable occupation_mother
v5 = c(22:30)
data = mca1_vars_df[v5,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

Observant la variable *Occupation_mother* es veu que totes les modalitats d'aquesta variable s'agrupen al voltant del centre. Si bé hi ha algunes variables que es poden situar de forma més clara en un dels quadrants o lluny de l'origen, en quedar totes molt a prop entre elles no hi ha prou certesa per assegurar que aporten informació verídica. 

```{r, out.width="90%"}
# variable course
v6 = c(8:15)
data = mca1_vars_df[v6,]

ggplot(data = data, aes(x = Dim.1, y = Dim.2)) +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") +
  geom_point(colour = "gray50", alpha = 0.7) +
  geom_density2d(colour = "gray80") +
  geom_text(data = data, 
            aes(x = Dim.1, y = Dim.2, 
                label = rownames(data), colour = Variable), cex = 3) +
  ggtitle("MCA plot of variables using R package FactoMineR") +
  scale_colour_discrete(name = "Variable") + 
  xlim(-1,2) + 
  ylim(-1.5,2)
```

Pel que fa a la variable *Course*, es veu que gran part de les seves modalitats s'agrupen al centre menys una. La categoria 'Management(PM)' se situa lluny de totes les altres i en el quadrant I. D'altra banda, les modalitats 'Social Service' i 'Nursing' si situen també lleugerament lluny del centre i en el quadrant II.

## Associacions

Havent estudiat cada una de les variables per separat s'ha pogut veure que hi ha certes modalitats de diferents variables que se situen en una mateixa zona del gràfic. Això significa que pot haver-hi certa associació entre aquestes modalitats, tot i que no es pot entrar a definir quin tipus de relació es dona entre elles. També s'ha de tenir en compte que havent utilitzat aquestes dues dimensions només es pot explicar un percentatge reduït de la variabilitat (24.25%). Així doncs, si es tinguessin les eines per poder utilitzar més dimensions, es podrien obtenir uns resultats més verídics i lleugerament diferents.

Les associacions que s'han trobat són les següents:

- En el quadrant I hi coincideixen aquells individus que van de tardes i els que estudien el curs de Management.

- En el quadrant II es troben aquells estudiants que estudien Social Service o Nursing, que han aconseguit graduar-se i han tingut una beca. 

- En el quadrant IV es troben generalment els estudiants homes i els que han abandonat la carrera.
