==> ML_UnsupervisedLearning_AcademicSuccess/08_clustering.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

# Clustering jeràrquic sobre les components factorials retingudes a l’ACP

Es fa el clustering per l'ACP. Tenim les coordenades de cada individu en les 5 dimensions que comprenen un 80% de la inèrcia. Amb aquestes dades es calcula la distància euclidiana, ja que són dades numèriques, s'aplica el mètode de Ward per fer el clustering, i posteriorment graficar el dendrograma. 

```{r, include=FALSE}
# install.packages("factoextra")
# install.packages("NbClust")
library(factoextra)
library(FactoMineR)
library('dendextend')
library(scales)
library(cluster)

load("data_post_pca.RData")
numeriques <- which(sapply(data,is.numeric))
numeriques

numer <- data[,numeriques]
sapply(numer,class)
pc1 <- prcomp(numer, scale = TRUE)
psi <- pc1$x[,1:5]
hc2 <- hclust(daisy(psi, metric = "euclidean"), "ward.D2")
dendo2 <- as.dendrogram(hc2)
```

Es veu clarament que el nombre òptim de grups per classificar les dades és 7:

```{r, out.width="80%"}
plot(dendo2)
```

Es visualitzen els 7 clústers:

```{r, out.width="80%"}
classes <- cutree(hc2, 7)
# save(classes, file="7classes.RData")
colors <- c("salmon","orange3","chartreuse4","lightseagreen","deepskyblue","lightslateblue","maroon3")
colors_dendo <- color_branches(dendo2, labels = classes, k = 7, col = colors)
plot(colors_dendo)
```
