==> ML_UnsupervisedLearning_AcademicSuccess/09_profiling.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Profiling post ACP i ACM

Cargamos los datos

```{r}
load("data_post_pca.RData")
load("7classes.RData")
```

Número de personas en cada grupo

```{r}
summary(factor(classes))
```

Definimos las funciones que calculan la significación de las variables mediante el p-valor

```{r}
ValorTestXnum <- function(Xnum,P){
  #freq dis of fac
  nk <- as.vector(table(P));
  n <- sum(nk);
  #mitjanes x grups
  xk <- tapply(Xnum,P,mean);
  #valors test
  txk <- (xk-mean(Xnum))/(sd(Xnum)*sqrt((n-nk)/(n*nk)));
  #p-values
  pxk <- pt(txk,n-1,lower.tail=F);
  for(c in 1:length(levels(as.factor(P)))){if (pxk[c]>0.5){pxk[c]<-1-pxk[c]}}
  return (pxk)
}


ValorTestXquali <- function(P,Xquali){
  taula <- table(P,Xquali);
  n <- sum(taula);
  pk <- apply(taula,1,sum)/n;
  pj <- apply(taula,2,sum)/n;
  pf <- taula/(n*pk);
  pjm <- matrix(data=pj,nrow=dim(pf)[1],ncol=dim(pf)[2], byrow=TRUE);
  dpf <- pf - pjm;
  dvt <- sqrt(((1-pk)/(n*pk))%*%t(pj*(1-pj)));
  zkj <- dpf
  zkj[dpf!=0]<-dpf[dpf!=0]/dvt[dpf!=0];
  pzkj <- pnorm(zkj,lower.tail=F);
  for(c in 1:length(levels(as.factor(P)))){for (s in 1:length(levels(Xquali))){if (pzkj[c,s]> 0.5){pzkj[c,s]<-1- pzkj[c,s]}}}
  return (list(rowpf=pf,vtest=zkj,pval=pzkj))
}
```

Preparamos los objetos necesarios para las herramientas gráficas

```{r}
#datos sin columna de clústers
dades <- data

#número de variables
K <- ncol(dades)

#grupos de los clústers
P <- classes
nameP <- "classe"

#número de clústers
nc <- length(levels(factor(P)))

#matriz y vector donde añadiremos los p-valores de cada variable
pvalk <- matrix(data=0,nrow=nc,ncol=K, dimnames=list(levels(as.factor(P)),names(dades)))
chis <- c()

#número de observaciones
n <- dim(dades)[1]
```

Calculamos los p-valores de las variables categóricas y numéricas

```{r}
# Generamos dos tablas, una para las categóricas y otra para las numéricas
for(k in 1:K){
  if (is.numeric(dades[,k])){
    for(s in levels(as.factor(P)))
    pvalk[,k]<-ValorTestXnum(dades[,k], P)
  }else{
    tab <- table(dades[,k],as.factor(P))
    chis[k]<-(chisq.test(tab)$p.value)
  }
}

#Añadimos los nombres de variable a las tablas
names(chis) <- colnames(pvalk)
# Juntamos las tablas
pvalorsvars <- colMeans(pvalk)
pvalorsvars[c(1:9,18)]<- chis[c(1:9,18)]

#Creamos una tabla que nos indica si es sigificativa o no
significativa <- pvalorsvars
for(i in 1:length(pvalorsvars)){
  if(pvalorsvars[i]<0.05){significativa[i]<-1}else{significativa[i] <- 0}
}
```

Tabla de los p-valores de las numéricas

```{r}
pvalk[,10:17]
```

Tabla de los p-valores de las categóricas

```{r}
chis[c(1:9,18)]
```

Tabla que nos dice si son significativas o no

SIGNIFICATIVA

Generación de todos los gráficos de las variables que son significativas

```{r}
for(k in 1:K){
  if (is.numeric(dades[,k]) && significativa[k]==1){
    print(paste("Anàlisi per classes de la Variable:", names(dades)[k]))
    barplot(tapply(dades[[k]], P, mean),main=paste("Means of", names(dades)[k], "by", nameP ))
    abline(h=mean(dades[[k]]))
    legend(0,mean(dades[[k]]),"global mean",bty="n")
    print("Estadístics per groups:")
    for(s in levels(as.factor(P))) {print(summary(dades[P==s,k]))}
    o<-oneway.test(dades[,k]~P)
    print(paste("p-valueANOVA:", o$p.value))
    kw<-kruskal.test(dades[,k]~P)
    print(paste("p-value Kruskal-Wallis:", kw$p.value))
    print("p-values ValorsTest: ")
    print(pvalk[,k])
  }else if(!is.numeric(dades[,k]) && significativa[k]==1){
      #qualitatives

      print(paste("Variable", names(dades)[k]))
      table<-table(P,dades[,k])
      #   print("Cross-table")
      #   print(table)
      rowperc<-prop.table(table,1)

      colperc<-prop.table(table,2)
      #  print("Distribucions condicionades a files")
      # print(rowperc)

      #ojo porque si la variable es true o false la identifica amb el tipus Logical i
      #aquest no te levels, por tanto, coertion preventiva

      dades[,k]<-as.factor(dades[,k])


      marg <- table(as.factor(P))/n
      print(append("Categories=",levels(as.factor(dades[,k]))))

      #from next plots, select one of them according to your practical case

      par(mar=c(5,4,4,8), xpd=TRUE)
      plot(marg,type="l",ylim=c(0,1),main=paste("Prop. of pos & neg by",names(dades)[k]))
      paleta<-rainbow(length(levels(dades[,k])))
      for(c in 1:length(levels(dades[,k]))){lines(colperc[,c],col=paleta[c]) }
      legend("topright", levels(dades[,k]), col=paleta, lty=2, cex=0.6)

      #condicionades a classes
      print(append("Categories=",levels(dades[,k])))
      par(mar=c(5,4,4,8), xpd=TRUE)
      plot(marg,type="n",ylim=c(0,1),main=paste("Prop. of pos & neg by",names(dades)[k]))
      paleta<-rainbow(length(levels(dades[,k])))
      for(c in 1:length(levels(dades[,k]))){lines(rowperc[,c],col=paleta[c]) }
      legend("topright", levels(dades[,k]), col=paleta, lty=2, cex=0.6)

      #amb variable en eix d'abcisses
      marg <-table(dades[,k])/n
      print(append("Categories=",levels(dades[,k])))
      par(mar=c(5,4,4,8), xpd=TRUE)
      plot(marg,type="l",ylim=c(0,1),main=paste("Prop. of pos & neg by",names(dades)[k]), las=3)
      #x<-plot(marg,type="l",ylim=c(0,1),main=paste("Prop. of pos & neg by",names(dades)[k]), xaxt="n")
      #text(x=x+.25, y=-1, adj=1, levels(CountryName), xpd=TRUE, srt=25, cex=0.7)
      paleta<-rainbow(length(levels(as.factor(P))))
      for(c in 1:length(levels(as.factor(P)))){lines(rowperc[c,],col=paleta[c]) }
      legend("topright", levels(as.factor(P)), col=paleta, lty=2, cex=0.6)

      #condicionades a columna
      par(mar=c(5,4,4,8), xpd=TRUE)
      plot(marg,type="n",ylim=c(0,1),main=paste("Prop. of pos & neg by",names(dades)[k]), las=3)
      paleta<-rainbow(length(levels(as.factor(P))))
      for(c in 1:length(levels(as.factor(P)))){lines(colperc[c,],col=paleta[c]) }
      legend("topright", levels(as.factor(P)), col=paleta, lty=2, cex=0.6)

      table<-table(dades[,k],P)
      print("Cross Table:")
      print(table)
      print("Distribucions condicionades a columnes:")
      print(colperc)

      #diagrames de barres adosades
      par(mar=c(5,4,4,8), xpd=TRUE)
      paleta<-rainbow(length(levels(dades[,k])))
      barplot(table(dades[,k], as.factor(P)), beside=TRUE,col=paleta)
      legend("topright",levels(as.factor(dades[,k])),pch=1,cex=0.5, col=paleta)

      print("Test Chi quadrat: ")
      print(chisq.test(dades[,k], as.factor(P)))

      print("valorsTest:")
      print( ValorTestXquali(P,dades[,k]))

      # g[k] <- ValorTestXquali(P,dades[,k])
      #calcular els pvalues de les quali
    }
  }#endfor
```
