==> ML_UnsupervisedLearning_AcademicSuccess/10_factorialPlanesWClusters.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
```

## Clústers representats sobre els dos primers plans factorials

```{r}
library(ggplot2)
data = get(load('prepro_data.Rdata'))
load("7classes.Rdata")
numeric <- which(sapply(data, is.numeric))
data_numeric <- data[, numeric]
pc1 <- prcomp(data_numeric, scale = TRUE)
```

Es representen les dades dividides per grups en els dos primers plans factorials de l'ACP:

```{r, out.width="95%"}
numer <- data_numeric
numer$clusters <- as.factor(classes)

PC1 <- pc1$x[,1]
PC2 <- pc1$x[,2]
ggplot(numer, 
       aes(x = PC1, 
           y = PC2, 
           color = clusters)) +
  geom_point() +
  stat_ellipse() +
  geom_hline(yintercept = 0, colour = "gray70") +
  geom_vline(xintercept = 0, colour = "gray70") 
```

Al gràfic es pot veure que per les dues primeres dimensions, que estan explicades especialment per les notes, els estudiants de veterinària (la segona component només) i els estudiants d'infermeria, els grups que s'allunyen del centre són els 4 i 7.

El grup 7 està conformat per les persones que estudien infermeria, tenen notes baixes, unitats curriculars suspeses i han deixat la carrera. El grup 4 està conformat per estudiants de veterinària que tenen unitats curriculars suspeses. L'únic altre grup que es distingeix una mica és l'1, que són per persones que deixen els estudis. La resta de grups estan barrejats al centre. 

```{r}
# ggplot(numer, 
#        aes(x = PC1, 
#            y = PC2, 
#            color = clusters)) +
#   geom_point() +
#   stat_ellipse() +
#   geom_hline(yintercept = 0, colour = "gray70") +
#   geom_vline(xintercept = 0, colour = "gray70") + 
#   xlim(-3,2) + 
#   ylim(-5,2)
```
